requirements.txt
# Core
numpy>=1.26.0

# Vector search
faiss-cpu>=1.8.0

# Together.ai SDK
together>=1.3.0

# Web app
streamlit>=1.31.0

# Resume parsing
pypdf>=4.0.0

# Optional: load .env locally (recommended for dev)
python-dotenv>=1.0.1

# Networking (usually installed already, but safe)
requests>=2.31.0